formifier module created. app working/eslint passing
